MyCapstone/requirements.txt
# Core
django==4.2.9
python-dotenv==1.0.1
str2bool==1.1

# UI
django-admin-material-dashboard==1.0.10

# Deployment
whitenoise==6.6.0
gunicorn==21.2.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.1.0
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
Django==4.2.9
django-cors-headers==4.4.0
django-filter==24.3
django-taggit==6.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Faker==28.4.1
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
mysqlclient==2.2.4
numpy==2.1.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
psycopg2==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
service-identity==24.1.0
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
str2bool==1.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
whitenoise==6.6.0
zope.interface==7.0.3

# psycopg2-binary
# mysqlclient
